Move Internal(Request|Response)Frame to internal + change type of recipient
