+++ kroxylicious/src/main/java/io/kroxylicious/proxy/internal/InternalResponseFrame.java
@@ -3,32 +3,34 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.kroxylicious.proxy.frame;
+package io.kroxylicious.proxy.internal;
 
 import java.util.Objects;
 
 import org.apache.kafka.common.message.ResponseHeaderData;
 import org.apache.kafka.common.protocol.ApiMessage;
 
+import io.kroxylicious.proxy.filter.KrpcFilter;
+import io.kroxylicious.proxy.frame.DecodedResponseFrame;
 import io.kroxylicious.proxy.future.ProxyPromise;
 
 public class InternalResponseFrame<B extends ApiMessage> extends DecodedResponseFrame<B> {
 
-    private final Object recipient;
+    private final KrpcFilter recipient;
 
     private final ProxyPromise<? extends Object> promise;
 
-    public InternalResponseFrame(Object recipient, ProxyPromise<?> promise, short apiVersion, int correlationId, ResponseHeaderData header, B body) {
+    public InternalResponseFrame(KrpcFilter recipient, ProxyPromise<?> promise, short apiVersion, int correlationId, ResponseHeaderData header, B body) {
         super(apiVersion, correlationId, header, body);
         this.recipient = Objects.requireNonNull(recipient);
         this.promise = Objects.requireNonNull(promise);
     }
 
-    public boolean isRecipient(Object candidate) {
+    public boolean isRecipient(KrpcFilter candidate) {
         return recipient != null && recipient.equals(candidate);
     }
 
-    public Object recipient() {
+    public KrpcFilter recipient() {
         return recipient;
     }
